third_party/WebKit/LayoutTests/editing/selection/skip-not-adjacent-contenteditable-expected.txt

Ensure that moving a selection does not skip adjacent contenteditable if the current is contentEditable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selection.focusNode is before
PASS selection.focusNode is before
PASS successfullyParsed is true

TEST COMPLETE

Also bug if moved to here
Bug if moved to here
Selection starts here, and should not go out from here
Bug if moved to here
Also bug if moved to here
